PRO/miApiJWT/requirements.txt
uvicorn[standard]
fastapi
PyJWT
python-multipart
